REF: SpaceTypes
